+++ modules/ocl/src/color.cpp
@@ -82,7 +82,7 @@ void RGB2Gray_caller(const oclMat &src, oclMat &dst, int bidx)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&bidx));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {src.cols, src.rows, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(src.cols), static_cast<size_t>(src.rows), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "RGB2Gray", gt, lt, args, -1, -1, build_options);
 }
 void Gray2RGB_caller(const oclMat &src, oclMat &dst)
@@ -97,7 +97,7 @@ void Gray2RGB_caller(const oclMat &src, oclMat &dst)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&dst.step));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {src.cols, src.rows, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(src.cols), static_cast<size_t>(src.rows), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "Gray2RGB", gt, lt, args, -1, -1, build_options);
 }
 void RGB2YUV_caller(const oclMat &src, oclMat &dst, int bidx)
@@ -115,7 +115,7 @@ void RGB2YUV_caller(const oclMat &src, oclMat &dst, int bidx)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&bidx));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {src.cols, src.rows, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(src.cols), static_cast<size_t>(src.rows), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "RGB2YUV", gt, lt, args, -1, -1, build_options);
 }
 void YUV2RGB_caller(const oclMat &src, oclMat &dst, int bidx)
@@ -133,7 +133,7 @@ void YUV2RGB_caller(const oclMat &src, oclMat &dst, int bidx)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&bidx));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {src.cols, src.rows, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(src.cols), static_cast<size_t>(src.rows), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "YUV2RGB", gt, lt, args, -1, -1, build_options);
 }
 void YUV2RGB_NV12_caller(const oclMat &src, oclMat &dst, int bidx)
@@ -151,7 +151,7 @@ void YUV2RGB_NV12_caller(const oclMat &src, oclMat &dst, int bidx)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&dst.rows));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {dst.cols / 2, dst.rows / 2, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(dst.cols / 2), static_cast<size_t>(dst.rows / 2), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "YUV2RGBA_NV12", gt, lt, args, -1, -1, build_options);
 }
 void RGB2YCrCb_caller(const oclMat &src, oclMat &dst, int bidx)
@@ -169,7 +169,7 @@ void RGB2YCrCb_caller(const oclMat &src, oclMat &dst, int bidx)
     args.push_back( std::make_pair( sizeof(cl_int) , (void *)&bidx));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&src.data));
     args.push_back( std::make_pair( sizeof(cl_mem) , (void *)&dst.data));
-    size_t gt[3] = {src.cols, src.rows, 1}, lt[3] = {16, 16, 1};
+    size_t gt[3] = {static_cast<size_t>(src.cols), static_cast<size_t>(src.rows), 1}, lt[3] = {16, 16, 1};
     openCLExecuteKernel(src.clCxt, &cvt_color, "RGB2YCrCb", gt, lt, args, -1, -1, build_options);
 }
 void cvtColor_caller(const oclMat &src, oclMat &dst, int code, int dcn)
